Read full commit
* Add Paths to Constants
* Add a function to check if a pack exists
* Add i18n for all strings
* Properly render all icons by defining a scale, width, and height

For some reason, using an AbortController of any kind causes none of the Assets to fetch. I thought this was because of the abort being whenever the text changes, but even in an independent useEffect with [] and the abort being in cleanup it still breaks. TO-DO: Fix the abort controller.