Use largest scale of icon defined in icon pack
This method doesn't assume that the icon pack defined all of the scales for an icon anymore, and finds the largest scale of an icon that is defined by the icon pack. If it reaches 1x and it still can't find the icon, then it gives up and uses the official icon.

* Image.source now also statefully updates when the original source changes.
* Allow checking if a pack exists asynchronously. This method should be used wherever it can, as it also updates the iconpack.installed array when it's called.
* Properly animate iconpack download finishing, adding a new i18n key with it
* Only allow the index of completed assets downloaded to increase. Beforehand all 20 promises would eagerly execute concurrently leading to the `completed` amount fluctuating within an error boundary of 20. This only allows the index to increase, as the promises resolve.