Read full commit
Currently bulk is slower in all cases because of the lack of lazy

This update makes nothing use bulk anymore, however this should be reverted when the lazy system eternal has been working on is finished.

I tested with and without bulk and the speeds are, on average, 300ms worse than without bulk, a big difference.